requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
pycryptodome==3.19.0

